gateway/requirements.txt
paho-mqtt
time
json
